Kleine Anpassungen

+++ backend/src/routes/participant.js
@@ -17,7 +17,7 @@ const createParticipant = (dbRow) => {
 // *** POST /api/Participant *********************************************************
 router.post('/', async (req, res) => {
   const { project_id, user_id } = req.body;
-  logger.debug(`API Participant -> Register Participation: Project ID: ${project_id}, User ID: ${user_id}`);
+  logger.debug(`API: POST /api/participant -> Project ID: ${project_id}, User ID: ${user_id}`);
 
   if (!project_id || !user_id) {
     return res.status(400).send(ErrorMsg.VALIDATION.MISSING_FIELDS);
@@ -43,7 +43,7 @@ router.post('/', async (req, res) => {
 router.put('/:id', authenticateToken, async (req, res) => {
   const { project_id, user_id } = req.body;
   const { id } = req.params;
-  logger.debug(`API Participant -> Update Participation: Project ID: ${project_id}, User ID: ${user_id}`);
+  logger.debug(`API: PUT  /api/participant/${id} -> Project ID: ${project_id}, User ID: ${user_id}`);
 
   if (!project_id || !user_id) {
     return res.status(400).send(ErrorMsg.VALIDATION.MISSING_FIELDS);
@@ -71,7 +71,7 @@ router.put('/:id', authenticateToken, async (req, res) => {
 // *** GET /api/Participant *********************************************************
 router.get('/:id', authenticateToken, async (req, res) => {
   const { id } = req.params;
-  logger.debug(`API Participant -> Get Participation (id): ${id}`);
+  logger.debug(`API: GET  /api/participant/${id}`);
 
   const result = await db_get(`SELECT * FROM Participant WHERE Participant.id = ?`, [id]);
   if (result.err) return res.status(500).send(ErrorMsg.SERVER.ERROR);
@@ -83,7 +83,7 @@ router.get('/:id', authenticateToken, async (req, res) => {
 // *** DELETE /api/Participant *********************************************************
 router.delete('/:id', authenticateToken, async (req, res) => {
   const { id } = req.params;
-  logger.debug(`API Participant -> Delete Participation (id): ${id}`);
+  logger.debug(`API: DEL  /api/participant/${id}`);
 
   result = await db_run('DELETE FROM Participant WHERE id = ?', [id]);
   if (result.err) {
@@ -98,7 +98,11 @@ router.delete('/:id', authenticateToken, async (req, res) => {
 // *** DELETE /api/Participant *********************************************************
 router.delete('/', authenticateToken, async (req, res) => {
   const { project_id, user_id } = req.body;
-  logger.debug(`API Participant -> Delete Participation: Project ID: ${project_id}, User ID: ${user_id}`);
+  logger.debug(`API: DEL  /api/participant -> Project ID: ${project_id}, User ID: ${user_id}`);
+
+  if (!project_id || !user_id) {
+    return res.status(400).send(ErrorMsg.VALIDATION.MISSING_FIELDS);
+  }
 
   result = await db_run('DELETE FROM Participant WHERE project_id = ? AND user_id = ?', [project_id, user_id]);
   if (result.err) {
